Change Id to Long

+++ src/main/java/com/example/todobackend/model/Category.java
@@ -14,7 +14,7 @@ public class Category {
     @Id
     @Column
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column
     private String title;
@@ -34,17 +34,17 @@ public class Category {
     public Category() {
     }
 
-    public Category(int id, String title, String description) {
+    public Category(Long id, String title, String description) {
         this.id = id;
         this.title = title;
         this.description = description;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
