+++ src/main/java/com/example/todobackend/model/Item.java
@@ -11,7 +11,7 @@ public class Item {
     @Id
     @Column
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column
     private String title;
@@ -32,17 +32,17 @@ public class Item {
     public Item() {
     }
 
-    public Item(int id, String title, String note) {
+    public Item(Long id, String title, String note) {
         this.id = id;
         this.title = title;
         this.note = note;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
