+++ src/main/java/com/example/todobackend/model/User.java
@@ -13,7 +13,7 @@ public class User {
     @Id
     @Column
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
 
     @Column
     private String firstName;
@@ -35,7 +35,7 @@ public class User {
     public User() {
     }
 
-    public User(int id, String firstName, String lastName, String emailAddress, String password) {
+    public User(Long id, String firstName, String lastName, String emailAddress, String password) {
         this.id = id;
         this.firstName = firstName;
         this.lastName = lastName;
@@ -43,11 +43,11 @@ public class User {
         this.password = password;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
